@@ -1,5 +1,6 @@
 {
   "name": "poor-richard",
+  "type": "module",
   "version": "0.0.1",
   "description": "Static website for Spotlight PA",
   "repository": "git@github.com:spotlightpa/poor-richard.git",
@@ -8,8 +9,8 @@
   "scripts": {
     "test": "yarn run test:eslint; yarn run test:prettier; yarn run test:html",
     "format": "yarn run format:eslint; yarn run format:prettier; yarn run format:html",
-    "test:eslint": "eslint --ext .js --ignore-path .gitignore .",
-    "format:eslint": "eslint --fix --ext .js --ignore-path .gitignore .",
+    "test:eslint": "eslint .",
+    "format:eslint": "eslint --fix .",
     "test:prettier": "prettier --ignore-path .gitignore --check '**/*.{scss,css,js,yml}'",
     "format:prettier": "prettier --ignore-path .gitignore --write '**/*.{scss,css,js,yml}'",
     "test:html": "prettier --ignore-path prettier-html-ignore.rc --check 'layouts/**/*.html'",
@@ -24,36 +25,26 @@
     "@alpinejs/collapse": "^3.14.1",
     "@alpinejs/focus": "^3.12.3",
     "@alpinejs/intersect": "^3.12.3",
-    "@babel/cli": "^7.22.9",
-    "@babel/core": "^7.22.9",
-    "@babel/preset-env": "^7.22.9",
-    "@fortawesome/fontawesome-free": "^6.4.0",
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-regular-svg-icons": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fullhuman/postcss-purgecss": "^5.0.0",
+    "@fortawesome/fontawesome-free": "^7.1.0",
+    "@fortawesome/fontawesome-svg-core": "^7.1.0",
+    "@fortawesome/free-regular-svg-icons": "^7.1.0",
+    "@fortawesome/free-solid-svg-icons": "^7.1.0",
     "@newswire/frames": "^1.0.0",
+    "@tailwindcss/cli": "^4.1.17",
     "alpinejs": "^3.12.3",
-    "autoprefixer": "^10.4.14",
     "journalize": "^2.5.1",
-    "postcss": "^8.4.31",
-    "postcss-cli": "^11.0.0",
-    "purgecss": "^5.0.0",
-    "smoothscroll-polyfill": "^0.4.4",
-    "tailwindcss": "3.4.13"
+    "tailwindcss": "^4.1.17"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^8.48.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-node": "^11.1.0",
+    "@eslint/compat": "^2.0.0",
+    "@eslint/eslintrc": "^3.3.3",
+    "@eslint/js": "^9.39.1",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-standard": "^5.0.0",
+    "globals": "^16.5.0",
     "prettier": "^3.0.3",
     "prettier-plugin-go-template": "^0.0.15",
-    "prettier-plugin-tailwindcss": "^0.6.5"
+    "prettier-plugin-tailwindcss": "^0.7.2"
   }
 }
