@@ -1,4 +1,5 @@
-module.exports = {
+export default {
+  tailwindConfig: "tailwind.config.js",
   plugins: ["prettier-plugin-tailwindcss", "prettier-plugin-go-template"],
   overrides: [
     {
